refactor: Reorganize code into asm, decode, and riscv modules
