+++ wasm-riscv-online/src/riscv/mod.rs
@@ -1,0 +1,2 @@
+pub mod imm;
+pub mod size;
