+++ wasm-riscv-online/src/riscv/size.rs
